Update lz4, add force option for downloader checksum.

+++ lib/downloader.rb
@@ -59,16 +59,21 @@ def downloader(url, sha256sum, filename = File.basename(url), verbose = false)
   calc_sha256sum = Digest::SHA256.hexdigest(File.read(filename))
 
   unless (sha256sum =~ /^SKIP$/i) || (calc_sha256sum == sha256sum)
-    FileUtils.rm_f filename
+    if @opt_force
+      puts "Updating checksum for #{filename} from #{sha256sum} to #{calc_sha256sum} in #{CREW_LOCAL_REPO_ROOT}/packages/#{@pkg_name}.rb .".lightblue
+      system "sed -i 's/#{sha256sum}/#{calc_sha256sum}/' #{CREW_LOCAL_REPO_ROOT}/packages/#{@pkg_name}.rb"
+    else
+      FileUtils.rm_f filename
 
-    warn 'Checksum mismatch :/ Try again?'.lightred, <<~EOT
-      #{''}
-                            Filename: #{filename.lightblue}
-          Expected checksum (SHA256): #{sha256sum.green}
-        Calculated checksum (SHA256): #{calc_sha256sum.red}
-    EOT
+      warn 'Checksum mismatch :/ Try again?'.lightred, <<~EOT
+        #{''}
+                              Filename: #{filename.lightblue}
+            Expected checksum (SHA256): #{sha256sum.green}
+          Calculated checksum (SHA256): #{calc_sha256sum.red}
+      EOT
 
-    exit 2
+      exit 2
+    end
   end
 rescue StandardError => e
   warn e.full_message
